Updated updateJobStatus and getNextJob services
updateJobStatus will be used to set jobs status to Failed or Finished by monitor script.
getNextJob now will check the vm currently accesses by direct url, and will try to find the next job to be run on that vm. If vm has jobs running already, does nothing, otherwise changes the status of next job to started and return the startString to monitor'